completed ex 4.21 - added authorization for delete

+++ controllers/blogs.js
@@ -16,7 +16,7 @@ blogRouter.post('/', async (request, response) => {
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
-  
+
   const user = await User.findById(decodedToken.id)
 
   if (!body.title || !body.url)
@@ -42,16 +42,40 @@ blogRouter.post('/', async (request, response) => {
   response.status(201).json(result)
 })
 
+// blogRouter.delete('/:id', async (request, response) => {
+//   id = request.params.id
+//   if (!mongoose.Types.ObjectId.isValid(id))
+//     return response.status(404).end()
+
+//   const result = await Blog.findByIdandDelAete(id)
+//   if (result)
+//     response.status(204).end()
+//   else
+//     response.status(404).end()
+// })
+
 blogRouter.delete('/:id', async (request, response) => {
   id = request.params.id
-  if (!mongoose.Types.ObjectId.isValid(id))
-    return response.status(404).end()
 
-  const result = await Blog.findByIdAndDelete(id)
-  if (result)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return response.status(404).end()
+  }
+
+  const blog = await Blog.findById(id)
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(id)
     response.status(204).end()
-  else
-    response.status(404).end()
+  }
+  else {
+    return response.status(401).json({ error: 'user unauthorized' })
+  }
 })
 
 blogRouter.put('/:id', async (request, response) => {
